+++ liquid/nodes.py
@@ -5,30 +5,40 @@ from functools import partial
 from contextlib import suppress
 from pathlib import Path
 import attr
-from .defaults import (LIQUID_COMPILED_RR_EXTEND,
-                       LIQUID_COMPILED_RR_APPEND,
-                       LIQUID_NODES,
-                       LIQUID_COMPILED_RENDERED,
-                       LIQUID_LOGLEVELID_DETAIL,
-                       LIQUID_LOGGER_NAME)
+from .defaults import (LIQUID_COMPILED_RR_EXTEND, LIQUID_COMPILED_RR_APPEND,
+                       LIQUID_NODES, LIQUID_COMPILED_RENDERED,
+                       LIQUID_LOGLEVELID_DETAIL, LIQUID_LOGGER_NAME)
 from .code import LiquidCode
 from .stream import safe_split
-from .exceptions import (LiquidSyntaxError,
-                         LiquidCodeTagExists,
+from .exceptions import (LiquidSyntaxError, LiquidCodeTagExists,
                          LiquidNodeAlreadyRegistered)
-
 LOGGER = logging.getLogger(LIQUID_LOGGER_NAME)
-
 # pylint: disable=no-value-for-parameter,fixme
 def register_node(name, klass):
-    """Register a new node"""
+    """@API
+    Register a new node
+    To unregister a node:
+    ```
+    from liquid.defaults import LIQUID_NODES
+    del LIQUID_NODES[<name>]
+    ```
+    @params:
+        name (string): The name of the node, will be match at `{% <name> ...`
+        klass (type): The class the handle the nodes"""
     if name in LIQUID_NODES:
         raise LiquidNodeAlreadyRegistered(f"Node {name!r} has already "
                                           "been registered")
     LIQUID_NODES[name] = klass
 
 def compact(string, left, right):
-    """Make string compact"""
+    """@API
+    Apply compact mode to the string
+    @params:
+        string (str): The string to apply compact mode
+        left (bool): If the left side of the string should be compacted
+        right (bool): If the right side of the string should be compacted
+    @returns:
+        (str): The string with compact mode applied"""
     if left:
         string = string.lstrip(" \t").lstrip('\n')
     if right:
@@ -36,7 +46,14 @@ def compact(string, left, right):
     return string
 
 def parse_mixed(mixed, shared_code):
-    """Parse mixed expressions like 1 + `a | @plus: 1`"""
+    """@API
+    Parse mixed expressions like 1 + `a | @plus: 1`
+    @params:
+        mixed (str): The mixed expression to be parsed
+        shared_code (LiquidCode): A LiquidCode object needed
+            by `NodeLiquidExpression`
+    @returns:
+        (str): The parsed expression"""
     from .expression import NodeLiquidExpression
     # safely split by backticks
     mixed = safe_split(mixed, '`', quotes="\"'", trim=False)
@@ -56,13 +73,18 @@ def parse_mixed(mixed, shared_code):
     return f"({ret})"
 
 def check_quotes(value):
-    """Check if value is correctly quoted or not quoted:
+    """@API
+    Check if value is correctly quoted or not quoted:
     True: <empty string>
     True: <unquoted string>
     False: 'abc <compond quote>
     False: "abc <compond quote>
     False: "abc' <compond quote>
-    False: 'abc" <compond quote>"""
+    False: 'abc" <compond quote>
+    @params:
+        value (string): The value to check
+    @returns:
+        (bool): True if value is correctly quoted otherwise False"""
     if len(value) == 1 and value in ('"', '\''):
         return False
     if ((value[:1] in ('"', '\'') or value[-1:] in ('"', '\'')) and
@@ -71,19 +93,29 @@ def check_quotes(value):
     return True
 
 def unquote(value):
-    """Remove quotes from a string
-    Assuming it's checked by check_quotes"""
+    """@API
+    Remove quotes from a string
+    Assuming it's checked by check_quotes
+    @params:
+        value (str): The value to be unquoted
+    @returns:
+        (str): The unquoted value"""
     return value[1:-1] if value[:1] in ('"', '\'') else value
 
 def scan_file(relpath, dirs, reverse=True):
-    """Scan a file in the dirs"""
+    """@API
+    Scan a file in the dirs
+    @params:
+        relpath (str): The relative path
+        dirs (list): The list of directories to scan
+        reverse (bool): Scan the last directory in the list first?
+    @returns:
+        (Path): The file found in the `dirs`"""
     relpath = Path(relpath)
     if relpath.is_absolute():
         return relpath
-
     if reverse:
         dirs = reversed(dirs)
-
     for directory in dirs:
         path = directory.joinpath(relpath)
         if path.is_file():
@@ -92,9 +124,17 @@ def scan_file(relpath, dirs, reverse=True):
 
 @attr.s(kw_only=True)
 class NodeVoid:
-    """Nodes without ending/closing
+    """@API
+    Nodes without ending/closing
     For example: {% config mode="compact" include="xxx" %}
-    """
+    @params:
+        name (str): The name of the node, will be matched at `{% <name>`
+        attrs (str): The rest string of the node other than name
+        code (LiquidCode): The LiquidCode object to save the codes
+        shared_code (LiquidCode): The LiquidCode object to save some
+            shared codes
+        config (LiquidConfig): The LiquidConfig object with the configurations
+        context (diot.Diot): The context of this node"""
     # The name of the node
     name = attr.ib()
     # mode="compact" include="xxx"
@@ -107,25 +147,31 @@ class NodeVoid:
     config = attr.ib(default=None, repr=False, eq=False)
     # the context, including filename, lineno, history, and stacks
     context = attr.ib(default=None, repr=False, eq=False)
-
     def __attrs_post_init__(self):
         if self.context:
             LOGGER.debug("[PARSER%2s] - Found node %r at %s:%s",
                          self.context.nstack, self.name,
                          self.context.filename, self.context.lineno)
-
     def start(self):
-        """Validate attrs"""
+        """@API
+        Node hit. Start validating the node and preparing for parsing"""
 
-    def _try_mixed(self, mixed):
-        """Try to parse mixed expression"""
+    def try_mixed(self, mixed):
+        """@API
+        Try to parse mixed expression using my shared_code
+        @params:
+            mixed (str): The mixed expression to be parsed
+        @returns:
+            (str): The parsed expression"""
         try:
             return parse_mixed(mixed, self.shared_code)
         except LiquidSyntaxError as lse:
             raise LiquidSyntaxError(str(lse), self.context) from lse
 
     def parse_node(self):
-        """Parse the node into python code"""
+        """@API
+        Parse the node into python codes, and add them to `self.code`
+        for execution"""
         if self.context:
             LOGGER.debug("[PARSER%2s]   Parsing node %r at %s:%s",
                          self.context.nstack, self.name,
@@ -134,19 +180,19 @@ class NodeVoid:
 
 @attr.s(kw_only=True)
 class Node(NodeVoid):
-    """Nodes need to be closed
-    For example: {% if ... %} ... {% endif %}
-    """
-
+    """@API
+    Nodes need to be closed
+    For example: {% if ... %} ... {% endif %}"""
     def parse_node(self):
-        """Returns False to parse content as literal
+        """@API
+        Returns False to parse content as literal
         Otherwise as normal liquid template"""
         super().parse_node()
         self.context.stacks.append(self)
 
     def end(self, name):
-        """End the node
-        """
+        """@API
+        End the node, check if I am the right end node to close."""
         last_node = self.context.stacks.pop(-1)
         if last_node.name != name:
             raise LiquidSyntaxError(f"Unmatched closing node 'end{name}' "
@@ -154,10 +200,19 @@ class Node(NodeVoid):
 
 @attr.s(kw_only=True)
 class NodeIntact(Node):
-    """Content intact node, where the content remains
+    """@API
+    Content intact node, where the content remains
     unparsed as liquid template.
     The content is grabbed between {% raw %} and {% endraw %}, which
-    will then be parsed in parse_content"""
+    will then be parsed in parse_content
+    @params:
+        content (str): The content between open and close node
+        lineno (int): The line number when the node hit, as `context.lineno`
+            will be changed during further parsing
+        compact_left (bool): Whether the left side of `content` should
+            be compact
+        compact_right (bool): Whether the right side of `content` should
+            be compact"""
     # we should grab content between open and close node
     content = attr.ib(default=None, init=False, eq=False, repr=False)
     # line number may be changed if parse is deferred
@@ -166,29 +221,25 @@ class NodeIntact(Node):
     compact_left = attr.ib(default=False, init=False, eq=False, repr=False)
     # should do compact on right of content?
     compact_right = attr.ib(default=False, init=False, eq=False, repr=False)
-
     def parse_content(self):
-        """Parse content between open and close nodes"""
+        """@API
+        Add debug information that I am parsing the content"""
         if self.context:
             LOGGER.debug("[PARSER%2s] - Parsing content of %snode %r at %s:%s",
                          self.context.nstack,
                          (f"deferred({self.defer}) "
                           if hasattr(self, "defer") else ""),
-                         self.name, self.context.filename,
-                         self.context.lineno)
+                         self.name, self.context.filename, self.context.lineno)
 
 @attr.s(kw_only=True)
 class NodeLiquidLiteral(NodeVoid):
-    """Any literals in the template"""
-
+    """@API
+    Any literals in the template"""
     name = attr.ib(default='<liquid_literal>')
-
     def parse_node(self):
         super().parse_node()
-
         if not self.attrs:
             return
-
         lines = self.attrs.splitlines(keepends=True)
         if len(lines) > 1:
             self.code.add_line(f"{LIQUID_COMPILED_RR_EXTEND}([")
@@ -203,8 +254,8 @@ class NodeLiquidLiteral(NodeVoid):
 
 @attr.s(kw_only=True)
 class NodeLiquidComment(NodeVoid):
-    """Node like {# ... #}"""
-
+    """@API
+    Node like {# ... #}"""
     name = attr.ib(default='<liquid_comment>')
     shared_code = attr.ib(default=None)
     code = attr.ib(default=None)
@@ -212,14 +263,12 @@ class NodeLiquidComment(NodeVoid):
     attrs = attr.ib(default='')
 
 class NodeConfig(NodeVoid):
-    """Node {% config mode="compact" include="../path/to" loglevel="info" %}
-    """
+    """@API
+    Node {% config mode="compact" include=".." loglevel="info" %}"""
     def start(self):
         configs = safe_split(self.attrs, ',')
-
         if not all(configs):
             raise LiquidSyntaxError("Empty configs found.", self.context)
-
         self.attrs = {}
         for config in configs:
             parts = safe_split(config, '=')
@@ -227,11 +276,9 @@ class NodeConfig(NodeVoid):
                 raise LiquidSyntaxError("Wrong format config [%s], "
                                         "expect 'key=\"value\"'" % config,
                                         self.context)
-
             if not check_quotes(parts[1]):
                 raise LiquidSyntaxError("Unclosed string in config: "
                                         f"{config}", self.context)
-
             self.attrs[parts[0]] = unquote(parts[1])
 
     def parse_node(self):
@@ -240,45 +287,46 @@ class NodeConfig(NodeVoid):
             self.config.__setattr__(key, val)
 
 class NodeIf(Node):
-    """{% if ... %}"""
+    """@API
+    Node {% if ... %} ... {% endif %}"""
     def start(self):
         # allow statement to end with ':'
         self.attrs = self.attrs.rstrip(':')
-
         if not self.attrs:
             raise LiquidSyntaxError(f"No expressions for {self.name!r} node",
                                     self.context)
-
     def parse_node(self):
         super().parse_node()
-        self.code.add_line(f"{self.name} {self._try_mixed(self.attrs)}:",
+        self.code.add_line(f"{self.name} {self.try_mixed(self.attrs)}:",
                            self.context)
         self.code.indent()
 
     def end(self, name):
         super().end(name)
+        # we need to add "pass" to support
+        # {% if 1 %}{% endif %}
+        if self.context.history[-1] is self:
+            self.code.add_line("pass")
         self.code.dedent()
 
 NodeWhile = NodeIf
 
 class NodeElse(NodeVoid):
-    """{% else %}"""
+    """@API
+    Node {% else %}"""
     def start(self):
         # allow statement to end with ':'
         self.attrs = self.attrs.rstrip(':')
         if self.attrs and self.attrs.split()[0] != "if":
             raise LiquidSyntaxError("No expressions allowed for 'else' node",
                                     self.context)
-
         if (not self.context.stacks or
                 self.context.stacks[-1].name not in ('case', 'if', 'unless')):
             raise LiquidSyntaxError("'else' must be in an "
                                     "'if/unless/case' node", self.context)
-
         if self.attrs == "if":
             raise LiquidSyntaxError("No expressions for 'else if' node",
                                     self.context)
-
         # if no statement found after if/unless
         # so that following is supported
         # {% if True %}{%else%} something else {% endif %}
@@ -295,28 +343,23 @@ class NodeElse(NodeVoid):
         else: # else if
             # we con't need to close this
             self.context.stacks.pop(-1)
-            NodeIf(name="if",
-                   code=self.code,
-                   shared_code=self.shared_code,
-                   context=self.context,
-                   config=self.config,
+            NodeIf(name="if", code=self.code, shared_code=self.shared_code,
+                   context=self.context, config=self.config,
                    attrs=self.attrs.split(maxsplit=1)[1]).parse_node()
 
 class NodeElseif(NodeVoid):
-    """{% else if %}"""
+    """@API
+    Node {% elseif .. %}, {% elif .. %} and {% elsif ... %}"""
     def start(self):
-
         self.attrs = self.attrs.rstrip(':')
         if not self.attrs:
             raise LiquidSyntaxError(f"No expressions for {self.name!r} node",
                                     self.context)
-
         if (not self.context.stacks or not isinstance(
                 self.context.stacks[-1], (NodeIf, NodeCase, NodeUnless)
         )):
             raise LiquidSyntaxError(f"{self.name!r} must be in an "
                                     "'if/unless/case' node", self.context)
-
         # if no statement found after if
         if (self.context.history and
                 isinstance(self.context.history[-1], NodeIf)):
@@ -325,32 +368,28 @@ class NodeElseif(NodeVoid):
 
     def parse_node(self):
         super().parse_node()
-        source = 'elif ' + self._try_mixed(self.attrs)
+        source = 'elif ' + self.try_mixed(self.attrs)
         self.code.add_line(source + ':', self.context)
         self.code.indent()
 
 class NodeRaw(NodeIntact):
-    """{% raw %} ... {% endraw %}"""
+    """@API
+    Node {% raw %} ... {% endraw %}"""
     def start(self):
         if self.attrs:
             raise LiquidSyntaxError("No expressions allowed for 'raw' node",
                                     self.context)
-
     def parse_content(self):
         super().parse_content()
         content = compact(self.content, self.compact_left, self.compact_right)
-
         NodeLiquidLiteral(attrs=content, code=self.code,
                           shared_code=self.shared_code,
                           context=self.context).parse_node()
 
-
 class NodeFor(Node):
-    """
-    Node '{% for ... %} {% endfor %}'
-    """
+    """@API
+    Node '{% for ... %} {% endfor %}'"""
     LIQUID_FORLOOP_CLASS = '_Liquid_forloop_class'
-
     def start(self):
         self.attrs = self.attrs.rstrip(':')
         parts = safe_split(self.attrs, ' in ', limit=1)
@@ -358,9 +397,8 @@ class NodeFor(Node):
             raise LiquidSyntaxError("'for' node expects format: "
                                     "'for var1, var2 in expr'", self.context)
 
-    def parse_node(self):
+    def parse_node(self): #pylint: disable=too-many-statements
         super().parse_node()
-
         with suppress(LiquidCodeTagExists), \
                 self.shared_code.tag(NodeFor.LIQUID_FORLOOP_CLASS) as tagged:
             tagged.add_line('')
@@ -422,17 +460,15 @@ class NodeFor(Node):
             tagged.add_line('return ret')
             tagged.dedent()
             tagged.dedent()
-
         # I am in stack already
         nest_fors = len([node for node in self.context.stacks
                          if isinstance(node, NodeFor)]) - 1
         if nest_fors > 0:
             # save my forloop object to avoid being overridden by inner ones
             self.code.add_line(f'forloop{nest_fors} = forloop')
-
         parts = safe_split(self.attrs, ' in ')
         # allow: for a, b in a + `1 | @increment`
-        parts[1] = self._try_mixed(parts[1])
+        parts[1] = self.try_mixed(parts[1])
         self.code.add_line(
             f'for forloop, {parts[0]} in '
             f'{NodeFor.LIQUID_FORLOOP_CLASS}({parts[1]}):',
@@ -450,9 +486,9 @@ class NodeFor(Node):
         self.code.dedent()
 
 class NodeCycle(NodeVoid):
-    """{% cycle ... %}"""
+    """@API
+    Node {% cycle ... %}"""
     def start(self):
-
         if not any(isinstance(node, (NodeWhile, NodeFor))
                    for node in self.context.stacks):
             raise LiquidSyntaxError("'cycle' node must be in a "
@@ -468,13 +504,11 @@ class NodeCycle(NodeVoid):
 
 @attr.s(kw_only=True)
 class NodeComment(Node):
-    """{% comment ... %} ... {% endcomment %}"""
-
+    """@API
+    Node {% comment ... %} ... {% endcomment %}"""
     LIQUID_COMMENT_PREFIX = '_liquid_comment'
     LIQUID_COMMENT_LINE_FORMATTER = '_liquid_comment_line_formatter'
-
     def start(self):
-
         self.attrs = self.attrs or '#'
         if len(self.attrs.split()) > 2:
             raise LiquidSyntaxError("Comments can only be wrapped by "
@@ -485,7 +519,6 @@ class NodeComment(Node):
 
     def parse_node(self):
         super().parse_node()
-
         with suppress(LiquidCodeTagExists), self.shared_code.tag(
                 NodeComment.LIQUID_COMMENT_LINE_FORMATTER
         ) as tagged:
@@ -501,11 +534,8 @@ class NodeComment(Node):
             tagged.dedent()
             tagged.add_line("return f'{prefix1} {line}{prefix2}{end}'")
             tagged.dedent()
-
-
         comcode = LiquidCode()
         self.context.parser.code.add_code(comcode)
-
         com_listname = f"{NodeComment.LIQUID_COMMENT_PREFIX}_{id(self)}"
         comcode.add_line("")
         comcode.add_line(f"# NODE COMMENT: {id(self)}")
@@ -542,8 +572,8 @@ class NodeComment(Node):
         self.context.parser.code = self.code
 
 class NodeBreak(NodeVoid):
-    """{% break %}"""
-
+    """@API
+    Node {% break %}"""
     def start(self):
         if self.attrs:
             raise LiquidSyntaxError("No expressions allowed for "
@@ -556,12 +586,11 @@ class NodeBreak(NodeVoid):
     def parse_node(self):
         super().parse_node()
         self.code.add_line(self.name, self.context)
-
 NodeContinue = NodeBreak
 
 class NodeAssign(NodeVoid):
-    """{% assign a = `1 | @plus: 1` %}"""
-
+    """@API
+    Node like {% assign a = `1 | @plus: 1` %}"""
     def start(self):
         parts = self.attrs.split('=', 1)
         if len(parts) < 2:
@@ -572,14 +601,13 @@ class NodeAssign(NodeVoid):
     def parse_node(self):
         super().parse_node()
         parts = self.attrs.split('=', 1)
-        values = self._try_mixed(parts[1].strip())
+        values = self.try_mixed(parts[1].strip())
         self.code.add_line(f"{parts[0].strip()} = {values}", self.context)
 
 class NodeCapture(Node):
-    """{% capture x %} ... {% endcapture %}"""
-
+    """@API
+    Node like {% capture x %} ... {% endcapture %}"""
     LIQUID_CAPTURE_PREFIX = '_liquid_capture'
-
     def start(self):
         if not self.attrs.isidentifier():
             raise LiquidSyntaxError("Not a valid variable name "
@@ -587,12 +615,11 @@ class NodeCapture(Node):
                                     self.context)
 
     def parse_node(self):
-        """Start a new code to save the content generated in the content"""
+        """@API
+        Start a new code to save the content generated in the content"""
         super().parse_node()
-
         capcode = LiquidCode()
         self.context.parser.code.add_code(capcode)
-
         capture_listname = f"{NodeCapture.LIQUID_CAPTURE_PREFIX}_{id(self)}"
         capcode.add_line("")
         capcode.add_line(f"# NODE CAPTURE: {id(self)}")
@@ -624,11 +651,11 @@ class NodeCapture(Node):
 
 @attr.s(kw_only=True)
 class NodeCase(Node):
-    """{% case x %} ... {% endcase %}"""
+    """@API
+    Node like {% case x %} ... {% endcase %}"""
     varname = attr.ib(default='', init=False, repr=False, eq=False)
     # when the first when hit, should use if, otherwise elif
     first_when = attr.ib(default=True, init=False, repr=False, eq=False)
-
     def start(self):
         if not self.attrs:
             raise LiquidSyntaxError(f"No values found for {self.name!r} node",
@@ -637,22 +664,22 @@ class NodeCase(Node):
     def parse_node(self):
         super().parse_node()
         self.varname = f"_case_{id(self)}"
-        value = self._try_mixed(self.attrs)
+        value = self.try_mixed(self.attrs)
         self.code.add_line(f"{self.varname} = {value}", self.context)
 
     def end(self, name):
         super().end(name)
         if self.first_when:
             raise LiquidSyntaxError(f"No 'when' node found in {self.name!r}")
+        self.code.dedent()
 
 class NodeWhen(NodeVoid):
-    """{% when ... %} in case"""
+    """@API
+    Node {% when ... %} in case"""
     def start(self):
-
         if not self.attrs:
             raise LiquidSyntaxError(f"No values found for {self.name!r} node",
                                     self.context)
-
         if (not self.context.stacks or
                 not isinstance(self.context.stacks[-1], NodeCase)):
             raise LiquidSyntaxError(f"{self.name!r} node must be in a "
@@ -660,24 +687,21 @@ class NodeWhen(NodeVoid):
 
     def parse_node(self):
         super().parse_node()
-
         varname = self.context.stacks[-1].varname
-        value = self._try_mixed(self.attrs)
+        value = self.try_mixed(self.attrs)
         equal = f"{varname} == {value}"
-
         if self.context.stacks[-1].first_when:
             self.code.add_line(f"if {equal}:", self.context)
-            self.code.indent()
             self.context.stacks[-1].first_when = False
         else:
             self.code.dedent()
             self.code.add_line(f"elif {equal}:", self.context)
-            self.code.indent()
+        self.code.indent()
 
 class NodeIncrement(NodeVoid):
-    """{% increment x %}"""
+    """@API
+    Node like {% increment x %}"""
     def start(self):
-
         if not self.attrs:
             raise LiquidSyntaxError("No variable specified for "
                                     f"{self.name!r} node", self.context)
@@ -690,13 +714,15 @@ class NodeIncrement(NodeVoid):
         self.code.add_line(f"{self.attrs} += 1", self.context)
 
 class NodeDecrement(NodeIncrement):
-    """{% increment y %}"""
+    """@API
+    Node like {% increment y %}"""
     def parse_node(self):
         NodeVoid.parse_node(self)
         self.code.add_line(f"{self.attrs} -= 1", self.context)
 
 class NodeFrom(NodeVoid):
-    """{% from ... import ... %}"""
+    """@API
+    Node {% from ... import ... %}"""
     def start(self):
         # pylint: disable=unsupported-membership-test
         if ' import ' not in self.attrs:
@@ -708,39 +734,39 @@ class NodeFrom(NodeVoid):
         self.code.add_line(f"from {self.attrs}", self.context)
 
 class NodeImport(NodeVoid):
-    """{% import ... %}"""
+    """@API
+    Node {% import ... %}"""
     def parse_node(self):
         super().parse_node()
         self.code.add_line(f"import {self.attrs}", self.context)
 
 class NodePython(Node):
-    """{% python a = 1 %} and {% python %}a = 1{% endpython %}"""
+    """@API
+    Node like {% python a = 1 %} and {% python %}a = 1{% endpython %}"""
     def parse_node(self):
         super().parse_node()
         if self.attrs:
             # it should be void, pop out of the stack
             self.context.stacks.pop(-1)
             self.code.add_line(self.attrs, self.context)
-        else:
-            ret = NodeIntact(name=self.name,
-                             attrs='',
-                             code=self.code,
-                             shared_code=self.shared_code,
-                             config=self.config,
-                             context=self.context)
-
-            ret.parse_content = partial(NodePython.parse_content, ret)
-            return ret
+            return None
+        ret = NodeIntact(name=self.name, attrs='', code=self.code,
+                         shared_code=self.shared_code, config=self.config,
+                         context=self.context)
+        ret.parse_content = partial(NodePython.parse_content, ret)
+        return ret
 
     def parse_content(self):
-        """Parse the content between {% python %} and {% endpython %}"""
+        """@API
+        Parse the content between {% python %} and {% endpython %}"""
         NodeIntact.parse_content(self)
         content = compact(self.content, self.compact_left, self.compact_right)
         for line in content.splitlines():
             self.code.add_line(line, self.context)
 
 class NodeUnless(Node):
-    """{% unless x %} ... {% endunless %}"""
+    """@API
+    Node like {% unless x %} ... {% endunless %}"""
     def start(self):
         self.attrs = self.attrs.rstrip(':')
         if not self.attrs:
@@ -749,8 +775,7 @@ class NodeUnless(Node):
 
     def parse_node(self):
         super().parse_node()
-
-        self.code.add_line(f"if not ({self._try_mixed(self.attrs)}):",
+        self.code.add_line(f"if not ({self.try_mixed(self.attrs)}):",
                            self.context)
         self.code.indent()
 
@@ -760,19 +785,15 @@ class NodeUnless(Node):
 
 @attr.s(kw_only=True)
 class NodeInclude(NodeVoid):
-    """{% include ... %}"""
-
+    """@API
+    Node {% include ... %}"""
     incfile = attr.ib(default=None, init=False)
     vars = attr.ib(default=attr.Factory(dict), init=False)
-
     LIQUID_INCLUDE_SOURCE = '_liquid_include_source'
-
     def start(self):
         if not self.attrs:
             raise LiquidSyntaxError("No file to include", self.context)
-
         parts = safe_split(self.attrs, " ", limit=1)
-
         if not check_quotes(parts[0]): # pragma: no cover
             # safe_split makes it impossible to Fail
             raise LiquidSyntaxError("Incorrectly quoted inclusion file: "
@@ -782,20 +803,17 @@ class NodeInclude(NodeVoid):
         incdirs = (self.config.include +
                    [Path(self.context.filename).resolve().parent])
         self.incfile = scan_file(parts[0], incdirs)
-
         incdirstr = ""
         if LOGGER.level < LIQUID_LOGLEVELID_DETAIL:
             incdirstr = "\nInclusion directories:\n"
             incdirstr += "\n".join(f"- {incdir}"
                                    for incdir in incdirs)
-
         if not self.incfile:
             raise LiquidSyntaxError("Cannot find file for inclusion: "
                                     f"{parts[0]}{incdirstr}", self.context)
         if not self.incfile.is_file():
             raise LiquidSyntaxError("File not exists for inclusion: "
                                     f"{parts[0]}{incdirstr}", self.context)
-
         if len(parts) > 1:
             vars_witheq = safe_split(parts[1], ',')
             for vareq in vars_witheq:
@@ -804,22 +822,22 @@ class NodeInclude(NodeVoid):
                                             f"{self.name!r} node", self.context)
             parts_eq = safe_split(vareq, '=', limit=1)
             if len(parts_eq) > 1:
-                self.vars[parts_eq[0]] = self._try_mixed(parts_eq[1])
+                self.vars[parts_eq[0]] = self.try_mixed(parts_eq[1])
             elif parts_eq[0].isidentifier():
                 self.vars[parts_eq[0]] = parts_eq[0]
             else:
                 raise LiquidSyntaxError("A variable or a kwarg needed "
                                         "for variables passed to "
                                         f"{self.name!r} node", self.context)
+        self.vars[LIQUID_COMPILED_RR_APPEND] = LIQUID_COMPILED_RR_APPEND
+        self.vars[LIQUID_COMPILED_RR_EXTEND] = LIQUID_COMPILED_RR_EXTEND
 
     def parse_node(self):
         super().parse_node()
-
         funcname = f"{NodeInclude.LIQUID_INCLUDE_SOURCE}_{id(self)}"
         varnames = ", ".join(self.vars)
         kwargs = ", ".join(f"{key}={val}" for key, val in self.vars.items())
         inccode = LiquidCode()
-
         self.code.add_line('')
         self.code.add_line(f"def {funcname}({varnames}):")
         self.code.indent()
@@ -829,28 +847,24 @@ class NodeInclude(NodeVoid):
         self.code.add_line('')
         self.code.dedent()
         self.code.add_line(f"{funcname}({kwargs})")
-
-        parser = self.context.parser.__class__(stream=None,
-                                               code=inccode,
-                                               shared_code=self.shared_code,
-                                               filename=self.incfile,
-                                               prev=(self.context.lineno,
-                                                     self.context.parser),
-                                               config=self.config)
-        parser.parse()
+        with self.config.tear() as teared_config:
+            parser = self.context.parser.__class__(
+                stream=None, code=inccode,
+                shared_code=self.shared_code, filename=self.incfile,
+                prev=(self.context.lineno, self.context.parser),
+                config=teared_config)
+            parser.parse()
 
 @attr.s(kw_only=True)
 class NodeBlock(Node):
-    """{% block ... %} {% endblock %}"""
-
+    """@API
+    Node {% block ... %} {% endblock %}"""
     def start(self):
-
         if not self.attrs:
             raise LiquidSyntaxError("A block name is required", self.context)
         if not self.attrs.isidentifier():
             raise LiquidSyntaxError(f"Not a valid block name: {self.attrs!r}",
                                     self.context)
-
     def parse_node(self):
         """Replace the parser's code with mine,
         so that the node inside me can be added to my code"""
@@ -878,29 +892,25 @@ class NodeBlock(Node):
 
 @attr.s(kw_only=True)
 class NodeExtends(NodeVoid):
-    """{% extends ... %}"""
+    """@API
+    Node {% extends ... %}"""
     # make sure extends parsed before blocks
     # because we have to choose the right blocks
     defer = attr.ib(default=999, init=False, eq=False)
     extfile = attr.ib(default=None, init=False)
     lineno = attr.ib(default=0, init=False)
-
     def start(self):
         self.lineno = self.context.lineno
-
         if not self.attrs:
             raise LiquidSyntaxError("No file to extend", self.context)
-
         if not check_quotes(self.attrs):# pragma: no cover
             # safe_split makes it impossible to Fail
             raise LiquidSyntaxError("Incorrectly quoted extension file: "
                                     f"{self.attrs!r}", self.context)
         self.attrs = unquote(self.attrs)
-
         if any(isinstance(node, NodeExtends) for node in self.context.history):
             raise LiquidSyntaxError(f"Only one {self.name!r} node allowed"
                                     " in a template", self.context)
-
         # also scan the directory with current template
         extends_dirs = (self.config.extends +
                         [Path(self.context.filename).resolve().parent])
@@ -908,8 +918,7 @@ class NodeExtends(NodeVoid):
         extdirstr = ""
         if LOGGER.level < LIQUID_LOGLEVELID_DETAIL:
             extdirstr = "\nExtension directories:\n"
-            extdirstr += "\n".join(f"- {extdir}"
-                                   for extdir in extends_dirs)
+            extdirstr += "\n".join(f"- {extdir}" for extdir in extends_dirs)
 
         if not self.extfile:
             raise LiquidSyntaxError("Cannot find file for extension: "
@@ -917,24 +926,22 @@ class NodeExtends(NodeVoid):
         if not self.extfile.is_file():
             raise LiquidSyntaxError("File not exists for extension: "
                                     f"{self.attrs}{extdirstr}", self.context)
-
     def parse_content(self):
         """Parse the mother template and impose my blocks into it
         And then replace the code of my parser with its."""
         # revert the lineno, as this is deferred
         self.context.lineno = self.lineno
         NodeIntact.parse_content(self)
-
-        parser = self.context.parser.__class__(
-            stream=None,
-            shared_code=self.shared_code,
-            code=LiquidCode(indent=self.code.ndent),
-            config=self.config,
-            filename=self.extfile,
-            prev=(self.context.lineno, self.context.parser)
-        )
-        parser.parse()
-
+        with self.config.tear() as teared_config:
+            parser = self.context.parser.__class__(
+                stream=None,
+                shared_code=self.shared_code,
+                code=LiquidCode(indent=self.code.ndent),
+                config=teared_config,
+                filename=self.extfile,
+                prev=(self.context.lineno, self.context.parser)
+            )
+            parser.parse()
         # we need to replace the core stuff of current parser with
         # those of parser, so that the assembler will assemble the code
         # from parser
@@ -947,18 +954,15 @@ class NodeExtends(NodeVoid):
                              self.context.nstack, bname, parser.context.nstack)
                 mother_code, index = mother_blocks[bname]
                 mother_code.codes[index] = binfo[0].codes[binfo[1]]
-
         # replace the code
         # what should we do with the shared code?
         curr_parser.code = parser.code
-
 
 class NodeMode(NodeVoid):
     """Deprecated {% mode ... %}"""
     def start(self):
         warnings.warn("'mode' node is deprecated, use 'config' node instead",
                       DeprecationWarning)
-
     def parse_node(self):
         super().parse_node()
         self.config.mode = self.attrs